api/requirements.txt
# Azure Functions for Python
azure-functions>=1.18.0

# Azure AI Services
azure-identity>=1.15.0
requests>=2.31.0

# Additional packages
python-dateutil>=2.8.2
python-multipart>=0.0.6
